# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLSEA Solar Generation API",
    "version": "0.1.0",
    "description": "SLSEA Solar Generation API, Stages 1\u20136. Public health and OpenAPI; JWT-protected hierarchy, immutable ingestion, scoped history and operational reads. Current paths are relative to the host root. GET/HEAD business representations support authorized ETag revalidation. Deployment and district summaries remain planned."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current API host"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/provinces": {
      "get": {
        "operationId": "listProvinces",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
    "/provinces/{provinceId}": {
      "get": {
        "operationId": "getProvince",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "provinceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": true
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
    "/provinces/{provinceId}/districts": {
      "get": {
        "operationId": "listProvinceDistricts",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "provinceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
    "/districts/{districtId}": {
      "get": {
        "operationId": "getDistrict",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": true
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
    "/districts/{districtId}/substations": {
      "get": {
        "operationId": "listDistrictSubstations",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
    "/substations/{substationId}": {
      "get": {
        "operationId": "getSubstation",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "substationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resource representation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "additionalProperties": true
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "description": " GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock."
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": {
        "description": "A server-generated UUID for correlating this response with application logs. For errors, it matches error.requestId.",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "7d50a2a4-39a5-46cc-a8ae-a6ccba1d3897"
      },
...
      "RepresentationTag": {
        "description": "Strong SHA-256 validator of the serialized representation. Mutable views include URL and jurisdiction context. Authorization is always checked before preconditions.",
        "schema": {
          "type": "string",
          "pattern": "^\"[a-f0-9]{64}\"$"
        }
      },
...
      "PrivateRevalidation": {
        "description": "Private caches must revalidate authenticated representations.",
        "schema": {
          "type": "string",
          "const": "private, no-cache"
        }
      }
    },
    "schemas": {
...
      "Collection": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid or expired credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string",
              "const": "Bearer"
            },
            "description": "Bearer authentication challenge."
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated principal lacks the required scope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "PreconditionFailed": {
        "description": "41201: a strong If-Match or applicable If-Unmodified-Since precondition failed after authentication, authorization and resource lookup.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived HS256 token issued by POST /auth/token."
      }
    }
  }
}